requirements.txt
openpyxl==3.0.9
python-dotenv==0.19.2
requests==2.27.1